LPS-85203 Rename and move.
